refactor(forge-host): consolidate scattered GitHub dirs under forge-host/github/ (R6, Task 6)
Wave 4 of the forge-host-adapter spec — the final consolidation:

Moved into forge-host/github/:
- github-accelerator-measurement/ → accelerator-measure.ts/.test.ts
- refresh-github-worldview/refresh.ts → removed (superseded by the
  full refresh-worldview.ts already in forge-host/github/)
- pr-preservation/README.md → pr-preservation-README.md (code already
  at pr-archive.ts from prior merge)

The github/ and archive/ directories were already consolidated in
PR #8060 (Wave 2 merge brought them in).

Fixed import paths:
- batch-resolve-pr-threads.ts: ai-attribution now local (./ai-attribution)
- poll-pr-gate-batch.ts: ferry-throttler + observe paths adjusted
  (../../ X instead of ../X, one level deeper now)
- accelerator-measure.test.ts: ./measure → ./accelerator-measure

Removed empty directories: github-accelerator-measurement/,
refresh-github-worldview/, pr-preservation/.

Updated .github/copilot-instructions.md path reference.

110 tests pass. Zero TS errors.
