+++ .github/copilot-instructions.md
@@ -343,7 +343,7 @@ Per [docs/CONFLICT-RESOLUTION.md](../docs/CONFLICT-RESOLUTION.md)
   - `docs/aurora/**` — courier-ferry archive
   - `docs/lost-substrate/**` — recovery and redundancy artifacts whose job is to preserve what was at risk of being lost
   - `docs/pr-preservation/**` — PR conversation archive
-  - `docs/pr-discussions/**` — PR discussion archive (generated by `src/Core.TypeScript/pr-preservation/archive-pr.ts`; preserves author attribution as provenance metadata)
+  - `docs/pr-discussions/**` — PR discussion archive (generated by `src/Core.TypeScript/forge-host/github/pr-archive.ts`; preserves author attribution as provenance metadata)
   - `docs/hygiene-history/**` — tick-history + drain-logs
   - `docs/active-trajectory.md` — load-state file + cumulative trajectory log; names + filenames-with-names are first-class here because the file's job is to preserve who-said-what + per-file named evidence
   - `docs/launch/**` — launch substrate (public-facing positioning artifacts; persona names + external creator attributions allowed because the substrate's job is to preserve the multi-agent factory's named-team positioning AND IP-respect attribution)
