+++ src/Core.TypeScript/forge-host/github/fixtures/github-status-pr-degraded.json
