+++ src/Core.TypeScript/hygiene/check-bash-retirement-inventory.ts
@@ -90,6 +90,7 @@ export const EXPECTED_RETAINED_SHELL: readonly string[] = [
   "full-ai-cluster/usb-nixos-installer/zeta-install.sh",
   "src/Core.TypeScript/kiro/kiro-loop-wrapper.sh",
   "src/Core.TypeScript/kiro/launchd/install.sh",
+  "tools/lint/lint-go-python.sh",
   "tools/setup/common/agent-clis.sh",
   "tools/setup/common/curl-fetch.sh",
   "tools/setup/common/dotnet-tools.sh",
@@ -139,6 +140,7 @@ export const RETAINED_SHELL_CATEGORY_BY_FILE: Readonly<Record<string, RetainedSh
   "full-ai-cluster/usb-nixos-installer/zeta-install.sh": "nixos installer",
   "src/Core.TypeScript/kiro/kiro-loop-wrapper.sh": "kiro loop wrapper",
   "src/Core.TypeScript/kiro/launchd/install.sh": "launchd bootstrap",
+  "tools/lint/lint-go-python.sh": "setup/bootstrap",
   "tools/setup/common/agent-clis.sh": "setup/bootstrap",
   "tools/setup/common/curl-fetch.sh": "setup/bootstrap",
   "tools/setup/common/dotnet-tools.sh": "setup/bootstrap",
